Moved implementation of functions to the correct source file.
